refactor: 예약 관련 Enum 추가
